+++ changelog.md
@@ -7,6 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- [Add distinct error types](https://github.com/ballerina-platform/ballerina-library/issues/8597)
+- [Incorporate the csv fail safe support in the FTP listener](https://github.com/ballerina-platform/ballerina-library/issues/8502)
+
+### Fixed
+
+
+## [2.16.0] - 2025-12-19
+
+- [Add close API for the FTP clients](https://github.com/ballerina-platform/module-ballerina-ftp/pull/1519)
 - [Added FTPS with SSL/TLS configuration](https://github.com/ballerina-platform/module-ballerina-ftp/pull/1523)
 - [Added support for advanced file selection and processing](https://github.com/ballerina-platform/ballerina-library/issues/8376)
 - [Improving Ballerina FTP Client with Copy, Move and IsExist API](https://github.com/ballerina-platform/ballerina-library/issues/8438)